Moved XML code to visualizations
